Refactor `XtbStreamConnection` and implement message building methods
The `XtbStreamConnection` interface is refactored by adjusting argument types for `subscribe` and `unsubscribe` methods and implementing additional methods for building and sending messages. Argument value must now be `Value::Object` and can be optional. A helper function `prepare_arguments` is added to validate and prepare arguments for the message. Also, an error type `InvalidArgumentsType` is included to handle inappropriate argument types.